Refactor app loading
