+++ lib/resolver.js
